refactor: 당첨 번호 입력 받기 기능 리팩토링 - WinnerNumber에서 검증 후에 당첨 번호 객체 생성 - WinnerNumber 테스트 작성g
